+++ CopilotKit/packages/react-core/src/components/copilot-provider/copilotkit-props.tsx
@@ -6,12 +6,12 @@ import { ReactNode } from "react";
 
 export interface CopilotKitProps {
   /**
-   * Your Copilot Cloud API key.
+   *  Your Copilot Cloud API key. Don't have it yet? <LinkToCopilotCloud>Click here to sign up for free</LinkToCopilotCloud>.
    */
   publicApiKey?: string;
 
   /**
-   * Cloud feature: Restrict input to a specific topic.
+   * Restrict input to a specific topic.
    */
   cloudRestrictToTopic?: {
     validTopics?: string[];
@@ -19,7 +19,7 @@ export interface CopilotKitProps {
   };
 
   /**
-   * the endpoint for the Copilot Runtime instance.
+   * The endpoint for the Copilot Runtime instance. [Click here for more information](/concepts/copilot-runtime).
    */
   runtimeUrl?: string;
 
@@ -34,32 +34,16 @@ export interface CopilotKitProps {
   textToSpeechUrl?: string;
 
   /**
-   * @deprecated use runtimeUrl instead
-   */
-  url?: string;
-
-  /**
    * Additional headers to be sent with the request.
    *
    * For example:
-   * ```js
+   * ```json
    * {
-   *   'Authorization': 'Bearer your_token_here'
+   *   "Authorization": "Bearer X"
    * }
    * ```
    */
   headers?: Record<string, string>;
-
-  /**
-   * Additional body params to be sent with the request
-   * For example:
-   * ```js
-   * {
-   *   'message': 'Hello, world!'
-   * }
-   * ```
-   */
-  body?: Record<string, any>;
 
   /**
    * The children to be rendered within the CopilotKit.
@@ -67,7 +51,7 @@ export interface CopilotKitProps {
   children: ReactNode;
 
   /**
-   * Backend only props that will be combined to body params to be sent with the request
+   * Custom properties to be sent with the request
    * For example:
    * ```js
    * {
@@ -75,5 +59,18 @@ export interface CopilotKitProps {
    * }
    * ```
    */
-  backendOnlyProps?: Record<string, any>;
+  properties?: Record<string, any>;
+
+  /**
+   * Indicates whether the user agent should send or receive cookies from the other domain
+   * in the case of cross-origin requests.
+   */
+  credentials?: RequestCredentials;
+
+  /**
+   * Whether to show the dev console.
+   *
+   * If set to "auto", the dev console will be show on localhost only.
+   */
+  showDevConsole?: boolean | "auto";
 }