+++ CopilotKit/packages/react-core/src/context/copilot-context.tsx
@@ -1,9 +1,5 @@
-import {
-  CopilotCloudConfig,
-  FunctionCallHandler,
-  Message,
-  ToolDefinition,
-} from "@copilotkit/shared";
+import { CopilotCloudConfig, FunctionCallHandler } from "@copilotkit/shared";
+import { Message } from "@copilotkit/runtime-client-gql";
 import { ActionRenderProps, FrontendAction } from "../types/frontend-action";
 import React from "react";
 import { TreeNodeId } from "../hooks/use-tree";
@@ -40,11 +36,6 @@ export interface CopilotApiConfig {
   textToSpeechUrl?: string;
 
   /**
-   * The endpoint for the chat API v2.
-   */
-  chatApiEndpointV2: string;
-
-  /**
    * additional headers to be sent with the request
    * @default {}
    * @example
@@ -57,19 +48,7 @@ export interface CopilotApiConfig {
   headers: Record<string, string>;
 
   /**
-   * Additional body params to be sent with the request
-   * @default {}
-   * @example
-   * ```
-   * {
-   *   'message': 'Hello, world!'
-   * }
-   * ```
-   */
-  body: Record<string, any>;
-
-  /**
-   * Backend only props that will be combined to body params to be sent with the request
+   * Custom properties to be sent with the request
    * @default {}
    * @example
    * ```
@@ -78,26 +57,27 @@ export interface CopilotApiConfig {
    * }
    * ```
    */
-  backendOnlyProps?: Record<string, any>;
+  properties?: Record<string, any>;
+
+  /**
+   * Indicates whether the user agent should send or receive cookies from the other domain
+   * in the case of cross-origin requests.
+   */
+  credentials?: RequestCredentials;
 }
 
 export type InChatRenderFunction = (props: ActionRenderProps<any>) => string | JSX.Element;
 
 export interface CopilotContextParams {
   // function-calling
-  entryPoints: Record<string, FrontendAction<any>>;
-  setEntryPoint: (id: string, entryPoint: FrontendAction<any>) => void;
-  removeEntryPoint: (id: string) => void;
+  actions: Record<string, FrontendAction<any>>;
+  setAction: (id: string, action: FrontendAction<any>) => void;
+  removeAction: (id: string) => void;
   chatComponentsCache: React.RefObject<Record<string, InChatRenderFunction | string>>;
-  getChatCompletionFunctionDescriptions: (
-    customEntryPoints?: Record<string, FrontendAction<any>>,
-  ) => ToolDefinition[];
+
   getFunctionCallHandler: (
     customEntryPoints?: Record<string, FrontendAction<any>>,
   ) => FunctionCallHandler;
-
-  messages: Message[];
-  setMessages: React.Dispatch<React.SetStateAction<Message[]>>;
 
   // text context
   addContext: (context: string, parentId?: string, categories?: string[]) => TreeNodeId;
@@ -109,7 +89,13 @@ export interface CopilotContextParams {
   removeDocumentContext: (documentId: string) => void;
   getDocumentsContext: (categories: string[]) => DocumentPointer[];
 
-  // suggestions configuration
+  // chat
+  messages: Message[];
+  setMessages: React.Dispatch<React.SetStateAction<Message[]>>;
+
+  isLoading: boolean;
+  setIsLoading: React.Dispatch<React.SetStateAction<boolean>>;
+
   chatSuggestionConfiguration: { [key: string]: CopilotChatSuggestionConfiguration };
   addChatSuggestionConfiguration: (
     id: string,
@@ -117,23 +103,36 @@ export interface CopilotContextParams {
   ) => void;
   removeChatSuggestionConfiguration: (id: string) => void;
 
+  chatInstructions: string;
+  setChatInstructions: React.Dispatch<React.SetStateAction<string>>;
+
   // api endpoints
   copilotApiConfig: CopilotApiConfig;
+
+  showDevConsole: boolean | "auto";
 }
 
 const emptyCopilotContext: CopilotContextParams = {
-  entryPoints: {},
-  setEntryPoint: () => {},
-  removeEntryPoint: () => {},
-  getChatCompletionFunctionDescriptions: () => returnAndThrowInDebug([]),
-  getFunctionCallHandler: () => returnAndThrowInDebug(async () => {}),
+  actions: {},
+  setAction: () => {},
+  removeAction: () => {},
+
   chatComponentsCache: { current: {} },
   getContextString: (documents: DocumentPointer[], categories: string[]) =>
     returnAndThrowInDebug(""),
   addContext: () => "",
   removeContext: () => {},
+
+  getFunctionCallHandler: () => returnAndThrowInDebug(async () => {}),
+
   messages: [],
   setMessages: () => returnAndThrowInDebug([]),
+
+  isLoading: false,
+  setIsLoading: () => returnAndThrowInDebug(false),
+
+  chatInstructions: "",
+  setChatInstructions: () => returnAndThrowInDebug(""),
 
   getDocumentsContext: (categories: string[]) => returnAndThrowInDebug([]),
   addDocumentContext: () => returnAndThrowInDebug(""),
@@ -143,9 +142,7 @@ const emptyCopilotContext: CopilotContextParams = {
     get chatApiEndpoint(): string {
       throw new Error("Remember to wrap your app in a `<CopilotKit> {...} </CopilotKit>` !!!");
     }
-    get chatApiEndpointV2(): string {
-      throw new Error("Remember to wrap your app in a `<CopilotKit> {...} </CopilotKit>` !!!");
-    }
+
     get headers(): Record<string, string> {
       return {};
     }
@@ -157,12 +154,17 @@ const emptyCopilotContext: CopilotContextParams = {
   chatSuggestionConfiguration: {},
   addChatSuggestionConfiguration: () => {},
   removeChatSuggestionConfiguration: () => {},
+  showDevConsole: "auto",
 };
 
 export const CopilotContext = React.createContext<CopilotContextParams>(emptyCopilotContext);
 
 export function useCopilotContext(): CopilotContextParams {
-  return React.useContext(CopilotContext);
+  const context = React.useContext(CopilotContext);
+  if (context === emptyCopilotContext) {
+    throw new Error("Remember to wrap your app in a `<CopilotKit> {...} </CopilotKit>` !!!");
+  }
+  return context;
 }
 
 function returnAndThrowInDebug<T>(value: T): T {