+++ CopilotKit/packages/react-core/src/hooks/use-copilot-action.ts
@@ -1,10 +1,9 @@
-import { useRef, useContext, useEffect } from "react";
+import { useRef, useEffect } from "react";
 import { FrontendAction } from "../types/frontend-action";
-import { CopilotContext } from "../context/copilot-context";
-import { nanoid } from "nanoid";
-import { Parameter } from "@copilotkit/shared";
+import { useCopilotContext } from "../context/copilot-context";
+import { Parameter, randomId } from "@copilotkit/shared";
 
-// We implement useCopilotActionImplementation dependency handling so that
+// We implement useCopilotAction dependency handling so that
 // the developer has the option to not provide any dependencies.
 // In this case, we assume they want to update the handler on each rerender.
 // To avoid getting stuck in an infinite loop, we update the handler directly,
@@ -17,16 +16,15 @@ export function useCopilotAction<const T extends Parameter[] | [] = []>(
   action: FrontendAction<T>,
   dependencies?: any[],
 ): void {
-  const { setEntryPoint, removeEntryPoint, entryPoints, chatComponentsCache } =
-    useContext(CopilotContext);
-  const idRef = useRef<string>(nanoid());
+  const { setAction, removeAction, actions, chatComponentsCache } = useCopilotContext();
+  const idRef = useRef<string>(randomId());
 
   // If the developer doesn't provide dependencies, we assume they want to
   // update handler and render function when the action object changes.
   // This ensures that any captured variables in the handler are up to date.
   if (dependencies === undefined) {
-    if (entryPoints[idRef.current]) {
-      entryPoints[idRef.current].handler = action.handler as any;
+    if (actions[idRef.current]) {
+      actions[idRef.current].handler = action.handler as any;
       if (typeof action.render === "function") {
         if (chatComponentsCache.current !== null) {
           chatComponentsCache.current[action.name] = action.render;
@@ -36,20 +34,24 @@ export function useCopilotAction<const T extends Parameter[] | [] = []>(
   }
 
   useEffect(() => {
-    setEntryPoint(idRef.current, action as any);
+    if (action.disabled) {
+      return;
+    }
+    setAction(idRef.current, action as any);
     if (chatComponentsCache.current !== null && action.render !== undefined) {
       chatComponentsCache.current[action.name] = action.render;
     }
     return () => {
       // NOTE: For now, we don't remove the chatComponentsCache entry when the action is removed.
       // This is because we currently don't have access to the messages array in CopilotContext.
-      removeEntryPoint(idRef.current);
+      removeAction(idRef.current);
     };
   }, [
-    setEntryPoint,
-    removeEntryPoint,
+    setAction,
+    removeAction,
     action.description,
     action.name,
+    action.disabled,
     // This should be faster than deep equality checking
     // In addition, all major JS engines guarantee the order of object keys
     JSON.stringify(action.parameters),