+++ CopilotKit/packages/react-core/src/hooks/use-copilot-chat.ts
@@ -1,83 +1,154 @@
-import { useMemo, useContext } from "react";
-import { CopilotContext } from "../context/copilot-context";
-import { Message, ToolDefinition } from "@copilotkit/shared";
+/**
+ * `useCopilotChat` is a React hook that lets you directly interact with the
+ * Copilot instance. Use to implement a fully custom UI (headless UI) or to
+ * programmatically interact with the Copilot instance managed by the default
+ * UI.
+ *
+ * ## Usage
+ *
+ * ### Simple Usage
+ *
+ * ```tsx
+ * import { useCopilotChat } from "@copilotkit/react-core";
+ * import { Role, TextMessage } from "@copilotkit/runtime-client-gql";
+ *
+ * export function YourComponent() {
+ *   const { appendMessage } = useCopilotChat();
+ *
+ *   appendMessage(
+ *     new TextMessage({
+ *       content: "Hello World",
+ *       role: Role.User,
+ *     }),
+ *   );
+ * }
+ * ```
+ *
+ * `useCopilotChat` returns an object with the following properties:
+ *
+ * ```tsx
+ * const {
+ *   visibleMessages, // An array of messages that are currently visible in the chat.
+ *   appendMessage, // A function to append a message to the chat.
+ *   setMessages, // A function to set the messages in the chat.
+ *   deleteMessage, // A function to delete a message from the chat.
+ *   reloadMessages, // A function to reload the messages from the API.
+ *   stopGeneration, // A function to stop the generation of the next message.
+ *   isLoading, // A boolean indicating if the chat is loading.
+ * } = useCopilotChat();
+ * ```
+ */
+import { useRef, useEffect, useCallback } from "react";
+import { useCopilotContext } from "../context/copilot-context";
+import { Message, Role, TextMessage } from "@copilotkit/runtime-client-gql";
 import { SystemMessageFunction } from "../types";
-import { UseChatOptions, useChat } from "./use-chat";
+import { useChat } from "./use-chat";
 import { defaultCopilotContextCategories } from "../components";
 
-export interface UseCopilotChatOptions extends UseChatOptions {
+export interface UseCopilotChatOptions {
+  /**
+   * A unique identifier for the chat. If not provided, a random one will be
+   * generated. When provided, the `useChat` hook with the same `id` will
+   * have shared states across components.
+   */
+  id?: string;
+
+  /**
+   * HTTP headers to be sent with the API request.
+   */
+  headers?: Record<string, string> | Headers;
+  /**
+   * System messages of the chat. Defaults to an empty array.
+   */
+  initialMessages?: Message[];
+
+  /**
+   * A function to generate the system message. Defaults to `defaultSystemMessage`.
+   */
   makeSystemMessage?: SystemMessageFunction;
-  additionalInstructions?: string;
 }
 
 export interface UseCopilotChatReturn {
   visibleMessages: Message[];
-  append: (message: Message) => Promise<void>;
-  reload: () => Promise<void>;
-  stop: () => void;
+  appendMessage: (message: Message) => Promise<void>;
+  setMessages: (messages: Message[]) => void;
+  deleteMessage: (messageId: string) => void;
+  reloadMessages: () => Promise<void>;
+  stopGeneration: () => void;
   isLoading: boolean;
-  input: string;
-  setInput: React.Dispatch<React.SetStateAction<string>>;
 }
 
 export function useCopilotChat({
   makeSystemMessage,
-  additionalInstructions,
   ...options
-}: UseCopilotChatOptions): UseCopilotChatReturn {
+}: UseCopilotChatOptions = {}): UseCopilotChatReturn {
   const {
     getContextString,
-    getChatCompletionFunctionDescriptions,
     getFunctionCallHandler,
     copilotApiConfig,
     messages,
     setMessages,
-  } = useContext(CopilotContext);
+    isLoading,
+    setIsLoading,
+    chatInstructions,
+    actions,
+  } = useCopilotContext();
 
-  const systemMessage: Message = useMemo(() => {
-    const systemMessageMaker = makeSystemMessage || defaultSystemMessage;
-    const contextString = getContextString([], defaultCopilotContextCategories); // TODO: make the context categories configurable
-
-    return {
-      id: "system",
-      content: systemMessageMaker(contextString, additionalInstructions),
-      role: "system",
-    };
-  }, [getContextString, makeSystemMessage, additionalInstructions]);
-
-  const functionDescriptions: ToolDefinition[] = useMemo(() => {
-    return getChatCompletionFunctionDescriptions();
-  }, [getChatCompletionFunctionDescriptions]);
-
-  const { append, reload, stop, isLoading, input, setInput } = useChat({
-    ...options,
-    copilotConfig: copilotApiConfig,
-    id: options.id,
-    initialMessages: [systemMessage].concat(options.initialMessages || []),
-    tools: functionDescriptions,
-    onFunctionCall: getFunctionCallHandler(),
-    headers: { ...options.headers },
-    body: {
-      ...options.body,
+  // We need to ensure that makeSystemMessageCallback always uses the latest
+  // useCopilotReadable data.
+  const latestGetContextString = useUpdatedRef(getContextString);
+  const deleteMessage = useCallback(
+    (messageId: string) => {
+      setMessages((prev) => prev.filter((message) => message.id !== messageId));
     },
-    messages,
-    setMessages,
-  });
-
-  const visibleMessages = messages.filter(
-    (message) =>
-      message.role === "user" || message.role === "assistant" || message.role === "function",
+    [setMessages],
   );
 
-  return {
-    visibleMessages,
-    append,
-    reload,
-    stop,
+  const makeSystemMessageCallback = useCallback(() => {
+    const systemMessageMaker = makeSystemMessage || defaultSystemMessage;
+    // this always gets the latest context string
+    const contextString = latestGetContextString.current([], defaultCopilotContextCategories); // TODO: make the context categories configurable
+
+    return new TextMessage({
+      content: systemMessageMaker(contextString, chatInstructions),
+      role: Role.System,
+    });
+  }, [getContextString, makeSystemMessage, chatInstructions]);
+
+  const { append, reload, stop } = useChat({
+    ...options,
+    actions: Object.values(actions),
+    copilotConfig: copilotApiConfig,
+    initialMessages: options.initialMessages || [],
+    onFunctionCall: getFunctionCallHandler(),
+    messages,
+    setMessages,
+    makeSystemMessageCallback,
     isLoading,
-    input,
-    setInput,
+    setIsLoading,
+  });
+
+  return {
+    visibleMessages: messages,
+    appendMessage: append,
+    setMessages,
+    reloadMessages: reload,
+    stopGeneration: stop,
+    deleteMessage,
+    isLoading,
   };
+}
+
+// store `value` in a ref and update
+// it whenever it changes.
+function useUpdatedRef<T>(value: T) {
+  const ref = useRef(value);
+
+  useEffect(() => {
+    ref.current = value;
+  }, [value]);
+
+  return ref;
 }
 
 export function defaultSystemMessage(