+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,5 +1,150 @@
 # @copilotkit/shared
 
+## 1.1.2
+
+### Patch Changes
+
+- Pin headless-ui/react version to v2.1.1
+
+## 1.1.1
+
+### Patch Changes
+
+- - improved documentation
+  - center textarea popup
+  - show/hide dev console
+  - forward maxTokens, stop and force function calling
+
+## 1.1.0
+
+### Minor Changes
+
+- Official support for Groq (`GroqAdapter`)
+
+## 1.0.9
+
+### Patch Changes
+
+- Dev console, bugfixes
+
+## 1.0.8
+
+### Patch Changes
+
+- Remove redundant console logs
+
+## 1.0.7
+
+### Patch Changes
+
+- Add \_copilotkit internal properties to runtime
+
+## 1.0.6
+
+### Patch Changes
+
+- - Proactively prevent race conditions
+  - Improve token counting performance
+
+## 1.0.5
+
+### Patch Changes
+
+- Include @copilotkit/runtime-client-gql NPM package version in request to Runtime
+
+## 1.0.4
+
+### Patch Changes
+
+- Remove nanoid
+
+## 1.0.3
+
+### Patch Changes
+
+- Add README.md to published packages and add keywords to package.json
+
+## 1.0.2
+
+### Patch Changes
+
+- Add README.md and homepage/url to published packages
+
+## 1.0.1
+
+### Patch Changes
+
+- Remove PostHog, use Segment Anonymous Telemetry instead
+
+## 1.0.0
+
+### Major Changes
+
+- b6a4b6eb: V1.0 Release Candidate
+
+  - A robust new protocol between the frontend and the Copilot Runtime
+  - Support for Copilot Cloud
+  - Generative UI
+  - Support for LangChain universal tool calling
+  - OpenAI assistant API streaming
+
+- V1.0 Release
+
+  - A robust new protocol between the frontend and the Copilot Runtime
+  - Support for Copilot Cloud
+  - Generative UI
+  - Support for LangChain universal tool calling
+  - OpenAI assistant API streaming
+
+### Patch Changes
+
+- b6a4b6eb: Introduce anonymous telemetry
+- b6a4b6eb: Set default Copilot Cloud runtime URL to versioned URL (v1)
+
+## 1.0.0-beta.2
+
+### Patch Changes
+
+- Set default Copilot Cloud runtime URL to versioned URL (v1)
+
+## 1.0.0-beta.1
+
+### Patch Changes
+
+- Introduce anonymous telemetry
+
+## 1.0.0-beta.0
+
+### Major Changes
+
+- V1.0 Release Candidate
+
+  - A robust new protocol between the frontend and the Copilot Runtime
+  - Support for Copilot Cloud
+  - Generative UI
+  - Support for LangChain universal tool calling
+  - OpenAI assistant API streaming
+
+## 0.37.0
+
+### Minor Changes
+
+- f771353: Fix: Stale CopilotReadable
+- 9df8d43: Remove unneeded tailwind components
+- CSS improvements, useCopilotChat, invisible messages
+
+## 0.37.0-mme-fix-textarea-css.1
+
+### Minor Changes
+
+- Remove unneeded tailwind components
+
+## 0.37.0-mme-fix-feedback-readable.0
+
+### Minor Changes
+
+- Fix: Stale CopilotReadable
+
 ## 0.36.0
 
 ### Minor Changes